test: updated tests all ok

+++ ruby/spec/find_by_spec.rb
@@ -7,7 +7,7 @@ describe 'Persistence Tests' do
       class Student
         include Persistence
         has_one String, named: :name
-        has_one Numeric, named: :age
+        has_one Integer, named: :age
         has_one Boolean, named: :is_regular
 
       end
