+++ ruby/spec/has_many_spec.rb
@@ -4,7 +4,7 @@ describe 'Persistence Tests' do
     before do
       class Grade
         include Persistence
-        has_one Numeric, named: :value
+        has_one Integer, named: :value
       end
       class Student
         include Persistence
